+++ changelog.md
@@ -2,5 +2,6 @@
 - added snap option for rotation and anchor position
 - added independent (free) rotation option
 - fixed button sprite scale issue (sprite size didn't match button touch box)
+- fixed issue when transform center sometimes initially isn't centered
 - fixed overlap of buttons with the controls after specific transforms
 - prevented zero-division crash when anchor is too close the control and you try to transform objects